add wrapper classes for embedding editor/event views
To resolve the problem with infinite embedding
in IEventView/IEditorView I created two classes
that should be used as a lazy-evalutation wrappers
around views that embed other views. This by itself
is not enough and you should *ALWAYS* choose a safe
default view when embedding other views (eg. HexView)
or choose no default view and leave it to user.